+++ src/database/src/Model/Relations/Concerns/AsPivot.php
@@ -199,24 +199,6 @@ trait AsPivot
     }
 
     /**
-     * Get the queueable identity for the entity.
-     */
-    public function getQueueableId()
-    {
-        if (isset($this->attributes[$this->getKeyName()])) {
-            return $this->getKey();
-        }
-
-        return sprintf(
-            '%s:%s:%s:%s',
-            $this->foreignKey,
-            $this->getAttribute($this->foreignKey),
-            $this->relatedKey,
-            $this->getAttribute($this->relatedKey)
-        );
-    }
-
-    /**
      * Get a new query to restore one or more models by their queueable IDs.
      *
      * @param array<int> $ids