+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 
 ## Fixed
 
+- [#361](https://github.com/hyperf-cloud/hyperf/pull/361) Fixed command `db:model` not work in mysql 8.
+
 # v1.0.10 - 2019-08-09
 
 ## Added