+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 
 - [#356](https://github.com/hyperf-cloud/hyperf/pull/356) Optimized aysnc-queue when push a job with `CodeGenerateInterface` param, it can be auto generate a small one.
 - [#358](https://github.com/hyperf-cloud/hyperf/pull/358) Write `annotations.cache` to file only enableCache.
+- [#359](https://github.com/hyperf-cloud/hyperf/pull/359) Added generate support for Collection and Model.
 
 # v1.0.10 - 2019-08-09