+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## Added
 
-- [#366](https://github.com/hyperf-cloud/hyperf/pull/366) Added InitProcessTitleListener, and `onStart`, `onManagerStart` events.
+- [#366](https://github.com/hyperf-cloud/hyperf/pull/366) Added `Hyperf\Server\Listener\InitProcessTitleListener` to init th process name, also added `Hyperf\Framework\Event\OnStart` and `Hyperf\Framework\Event\OnManagerStart` events.
 
 ## Fixed
 
@@ -14,7 +14,7 @@
 ## Changed
 
 - [#356](https://github.com/hyperf-cloud/hyperf/pull/356) Optimized aysnc-queue when push a job with `CodeGenerateInterface` param, it can be auto generate a small one.
-- [#358](https://github.com/hyperf-cloud/hyperf/pull/358) Write `annotations.cache` to file only enableCache.
+- [#358](https://github.com/hyperf-cloud/hyperf/pull/358) Only write the annotation cache file when `$enableCache` is `true`.
 - [#359](https://github.com/hyperf-cloud/hyperf/pull/359) Added generate support for Collection and Model.
 
 # v1.0.10 - 2019-08-09